refactor(path): update import path
